MYSQL must have UTC timezone.
By default, when running docker-compose up, the container for
MYSQL will have the timezone of the host (in my tests CEST).
That means that any connection against MYSQL will be considered by
the server as CEST connection.
As siptrace is using gettimeofday to fill the timestamp header in
HEPv2+, the timestamp received by homer-kamailio is in UTC.
When sipcapture inserts the packet into MYSQL, as the connection is
considered by the database server as CEST and the data type used is
timestamp, MYSQL will try to convert the inserted timestamp to UTC.
That means the inserted timestamp will be 2 hours behind the real
hour in my example.