Clear any pre-existing PID files
Now that we have a restart policy, the webapp container doesn't recover
as it should from a restart. This commit removes any previous PID files
so that Apache can start correctly. In theory this shouldn't be an issue
with docker due to it's design.

Per: https://github.com/docker-library/php/issues/187